Chore update scripts to 0.0.9

+++ src/lib/server/structs/universe.ts
@@ -1,10 +1,13 @@
 import { boolean, text } from 'drizzle-orm/pg-core';
-import { Struct } from 'drizzle-struct/back-end';
+import { Struct, StructStream } from 'drizzle-struct/back-end';
 import { attemptAsync, resolveAll } from 'ts-utils/check';
 import { Account } from './account';
 import { DB } from '../db';
-import { eq, sql } from 'drizzle-orm';
+import { and, eq, sql } from 'drizzle-orm';
 import { Permissions } from './permissions';
+import { Session } from './session';
+import { z } from 'zod';
+import { createEntitlement, readEntitlement } from '../utils/entitlements';
 
 export namespace Universes {
 	export const Universe = new Struct({
@@ -19,10 +22,10 @@ export namespace Universes {
 	Universe.on('delete', (u) => {
 		Struct.each((s) => {
 			s.each((d) => {
-				d.removeUniverses(u.id);
+				d.setUniverse('');
 			});
 
-			UniverseInvites.fromProperty('universe', u.id, {
+			UniverseInvite.fromProperty('universe', u.id, {
 				type: 'stream'
 			}).pipe((i) => i.delete());
 		});
@@ -30,22 +33,36 @@ export namespace Universes {
 
 	export type UniverseData = typeof Universe.sample;
 
-	export const getUniverses = async (account: Account.AccountData) => {
+	export const getUniverses = async (accountId: string) => {
 		return attemptAsync(async () => {
-			return resolveAll(
-				await Promise.all(
-					account
-						.getUniverses()
-						.unwrap()
-						.map((u) => Universes.Universe.fromId(u))
+			// return resolveAll(
+			// 	await Promise.all(
+			// 		account
+			// 			.getUniverses()
+			// 			.unwrap()
+			// 			.map((u) => Universes.Universe.fromId(u))
+			// 	)
+			// )
+			// 	.unwrap()
+			// 	.filter(Boolean) as Universes.UniverseData[];
+
+			const data = await DB.select()
+				.from(Universe.table)
+				.innerJoin(
+					Permissions.RoleAccount.table,
+					eq(Permissions.RoleAccount.table.account, accountId)
 				)
-			)
-				.unwrap()
-				.filter(Boolean) as Universes.UniverseData[];
+				.innerJoin(
+					Permissions.Role.table,
+					eq(Permissions.RoleAccount.table.role, Permissions.Role.table.id)
+				)
+				.where(eq(Permissions.Role.table.name, 'Member'));
+
+			return data.map((d) => Universe.Generator(d.universe));
 		});
 	};
 
-	export const UniverseInvites = new Struct({
+	export const UniverseInvite = new Struct({
 		name: 'universe_invite',
 		structure: {
 			universe: text('universe').notNull(),
@@ -54,7 +71,35 @@ export namespace Universes {
 		}
 	});
 
-	export type UniverseInviteData = typeof UniverseInvites.sample;
+	UniverseInvite.callListen('invite', async (event, data) => {
+		const session = (await Session.getSession(event)).unwrap();
+		const account = (await Session.getAccount(session)).unwrap();
+
+		if (!account) {
+			throw new Error('Not logged in');
+		}
+
+		const i = z
+			.object({
+				user: z.string(),
+				universe: z.string()
+			})
+			.parse(data);
+
+		const invitee = (await Account.Account.fromId(i.user)).unwrap();
+		if (!invitee) throw new Error('Account not found');
+
+		const u = (await Universe.fromId(i.universe)).unwrap();
+		if (!u) throw new Error('Universe not found');
+
+		await invite(u, invitee, account);
+
+		return {
+			success: true
+		};
+	});
+
+	export type UniverseInviteData = typeof UniverseInvite.sample;
 
 	export const createUniverse = async (
 		config: {
@@ -64,25 +109,32 @@ export namespace Universes {
 		},
 		account: Account.AccountData
 	) => {
-		attemptAsync(async () => {
+		return attemptAsync(async () => {
 			const u = (await Universe.new(config)).unwrap();
 			const admin = (
 				await Permissions.Role.new({
 					universe: u.id,
 					name: 'Admin',
-					permissions: '[*]',
-					description: `${u.data.name} Aministrator`
+					description: `${u.data.name} Aministrator`,
+					links: '[]',
+					entitlements: '[]'
 				})
 			).unwrap();
 			const member = (
 				await Permissions.Role.new({
 					universe: u.id,
 					name: 'Member',
-					permissions: '[]',
-					description: `${u.data.name} Member`
+					description: `${u.data.name} Member`,
+					links: '[]',
+					entitlements: '[]'
 				})
 			).unwrap();
-			(await account.addUniverses(u.id)).unwrap();
+			(
+				await admin.update({
+					entitlements: JSON.stringify(['manage-roles', 'manage-universe'])
+				})
+			).unwrap();
+			(await admin.setStatic(true)).unwrap();
 			await Permissions.RoleAccount.new({
 				role: admin.id,
 				account: account.id
@@ -93,6 +145,14 @@ export namespace Universes {
 					account: account.id
 				})
 			).unwrap();
+			(
+				await member.update({
+					entitlements: JSON.stringify(['view-roles', 'view-universe'])
+				})
+			).unwrap();
+			(await member.setStatic(true)).unwrap();
+
+			return u;
 		});
 	};
 
@@ -103,7 +163,7 @@ export namespace Universes {
 	) => {
 		return attemptAsync(async () => {
 			const invite = (
-				await UniverseInvites.new({
+				await UniverseInvite.new({
 					universe: universe.id,
 					account: account.id,
 					inviter: inviter.id
@@ -134,14 +194,14 @@ export namespace Universes {
 	) => {
 		return attemptAsync(async () => {
 			const res = await DB.select()
-				.from(UniverseInvites.table)
-				.innerJoin(Universe.table, eq(UniverseInvites.table.universe, Universe.table.id))
+				.from(UniverseInvite.table)
+				.innerJoin(Universe.table, eq(UniverseInvite.table.universe, Universe.table.id))
 				.limit(config.limit)
 				.offset(config.offset)
-				.where(sql`${UniverseInvites.table.account} = ${account.id}`);
+				.where(sql`${UniverseInvite.table.account} = ${account.id}`);
 
 			return res.map((r) => ({
-				invite: UniverseInvites.Generator(r.universe_invite),
+				invite: UniverseInvite.Generator(r.universe_invite),
 				universe: Universe.Generator(r.universe)
 			}));
 		});
@@ -153,7 +213,23 @@ export namespace Universes {
 
 			const a = await (await Account.Account.fromId(account)).unwrap();
 			if (!a) return;
-			(await a.addUniverses(universe)).unwrap();
+
+			const roles = (
+				await Permissions.Role.fromProperty('universe', universe, {
+					type: 'stream'
+				}).await()
+			).unwrap();
+
+			const member = roles.find((r) => r.data.name === 'Member'); // should always succeed because data is static
+
+			if (!member) throw new Error('Member role not found');
+
+			(
+				await Permissions.RoleAccount.new({
+					account: a.id,
+					role: member.id
+				})
+			).unwrap();
 
 			(await invite.delete()).unwrap();
 		});
@@ -162,7 +238,64 @@ export namespace Universes {
 	export const declineInvite = async (invite: UniverseInviteData) => {
 		return invite.delete();
 	};
+
+	export const isMember = async (
+		account: Account.AccountData,
+		universe: Universes.UniverseData
+	) => {
+		return attemptAsync(async () => {});
+	};
+
+	export const getMembers = async (universe: UniverseData) => {
+		return attemptAsync(async () => {
+			const data = await DB.select()
+				.from(Account.Account.table)
+				.innerJoin(
+					Permissions.RoleAccount.table,
+					eq(Account.Account.table.id, Permissions.RoleAccount.table.account)
+				)
+				.innerJoin(
+					Permissions.Role.table,
+					eq(Permissions.RoleAccount.table.role, Permissions.Role.table.id)
+				)
+				.where(eq(Permissions.Role.table.universe, universe.id));
+			return data
+				.filter((v, i, a) => a.findIndex((t) => t.account.id === v.account.id) === i)
+				.map((d) => Account.Account.Generator(d.account));
+		});
+	};
+
+	export const memberRoles = async (account: Account.AccountData, universe: UniverseData) => {
+		return attemptAsync(async () => {
+			const data = await DB.select()
+				.from(Permissions.RoleAccount.table)
+				.innerJoin(
+					Permissions.Role.table,
+					eq(Permissions.RoleAccount.table.role, Permissions.Role.table.id)
+				)
+				.where(
+					and(
+						eq(Permissions.RoleAccount.table.account, account.id),
+						eq(Permissions.Role.table.universe, universe.id)
+					)
+				);
+
+			return data.map((d) => Permissions.Role.Generator(d.role));
+		});
+	};
+
+	createEntitlement({
+		name: 'manage-universe',
+		struct: Universe,
+		permissions: ['*']
+	});
+
+	createEntitlement({
+		name: 'view-universe',
+		struct: Universe,
+		permissions: ['read:name', 'read:description', 'read:public']
+	});
 }
 
 export const _universeTable = Universes.Universe.table;
-export const _universeInviteTable = Universes.UniverseInvites.table;
+export const _universeInviteTable = Universes.UniverseInvite.table;
